[tests] Split up dev tests to increase concurrency
